Adding recent puzzle solutions.
There are a few that were super easy and now I am getting to a couple
that have some parts that I need to research things and try different
methods. There is no version control within the in-browser editor (and
the vim mode sucks), so I am incorporating this into my puzzle solving
workflow. I might have to change changes in how I test as well, but I
will have to solve that as I go I think.