Finished How Time Flies.
This one was interesting. I thought it would be a no-brainer easy puzzle
but it actually got me for a quick second because it was being very
exact about what constitutes a month (not just a default value like 30
or 31). I forgot that the dateutils library existed that does the
headache work of trying to do time deltas, so I was first attempting to
use the base datetime library and failing.