Finished Prefix code Puzzle.
Was just winding down for the day and wanted to complete an easy puzzle.
This one did not disappoint! I had done something similar with my Ultima
Online client where the bytestream would have different length packets
that could come bunched together. For that, I did something similar to
this where I had a dictionary of packet headers and would pass the
length of the given packet type to a packet handler and then skip over
that length to go on to the next packet dynamically.